requirements.txt
Pillow
pylast
